refactor: src/lib 直下の aiChat 系モジュールを src/lib/aiChat/ へ集約
aiChatActionHelpers / aiChatActions / aiChatConversationTitle /
aiChatDetailLandingPayload / aiChatPrompt / aiChatPromptActionFormat
（およびテスト）を移動し、AGENTS.md の配置規則
「同一ドメイン 2 ファイル以上はサブディレクトリへ」に従う。
import は @/ エイリアスへ統一して更新。

検証: ルート vitest 全 2652 テストがパス。
