+++ src/lib/aiChat/aiChatActionHelpers.test.ts
