+++ src/lib/aiChat/aiChatDetailLandingPayload.test.ts
